Interactive mode and other changes
When you run the script without arguments, the script will prompt you to specify the path for the srt and video files. In addition, the marker file will now include the subtitle text in the name field instead of the description field, allowing you to quickly glancce at the text on the timeline.